:new: Migrated request skript/ class to this project
The code handles a request to a website to get a table with covid numbers and generates a json file from the gathered data.

There are checks in place that the found data is actually from today before writing.

I also included a write of the raw request text, for the case that the structure of the website changes.

If the file it self gets exceuted. the script will try all two hours to get data and then fall asleep until the next day, whn data was found.