:new: Migrated analyzer class
This is also a migration of an existing script.
It builds ontop of the data gained by `requester.py`.
The files of the last 20 days will be processed to generate graphs with the current infections in each town.
Those graphs will be saved in a separate folder, where the bot will be able to access them.