:bug: Fixed crashes when sending to unauthorized chats
There was an error that the bot stopped requesting new data, when it tried to send data to a user who blocked the bot.

So the bot wen't silent from that moment.

The code now handles such chats.
There is just a try/ catch at the moment and no removal of that user from the database.

I also improved logging around the request structure to have more data when such an error occurs again.

And there is a reduction of the request period from 10 minutes to one hour.