:new: Added incidence calculation for each city
New function for getting the data of the last seven days
It checks if the last day might also contain the numbers of the previous day due to a missing day.
-> Chekcing via is_missing() function, which will check this edge case
If edge case exists: dividing 7th days data by number of missing days before to approximate the number of day 7

Then calculating the incidence.

The actual implementation of that number in the graph happended last commit by accident...