:new: Added RequesterV2 Class
This requester class connects to an other data source than it's predecessor.
It targets the continious press releases instead of the static site that wasn't updated on a daily basis.

The class checks for validity of the requested website and interates trough the releases until it finds a matching entry.
If nothing found the class will go to sleep