:arrow_upper_right: Removed old subscription system
Wow, it feels so good to get rid of that silly redundant command system, which contained zero flexibilty.
It was one of the first things I wrote for telegram and I didn't know it better back then.
But now I've replaced nine commands with just one!

Subscriptions are now inline only.