↗️ Changed data types from int to float
Reason: JSONs can now contain NaN values or miss a field. Pandas can represent this as float (NaN), but not as integer.
The whole module does only contain one multiplication / division which operate on summed values and the result is expected to be a float.
So there shouldn't be any problem doing that.