🆕 Added capture handling for different press releases
It's now possible to track two different press release layouts at once by changing the expected value-positions.
So its now possible to handle both release layouts.

This allows the capturing of the number of Britain mutations.
JSONs from old press releases will hold a field for B.1.1.7 which will be filled with NaN.
This also required an improvement of the logging reports.