🆕 Implement less static validations for better adjustments
The validation idea is still the same, but the patterns are no longer hardcoded as this AND this AND that.
They are now loaded from file and the system iterates trough a list that  contains all knows patterns.
If one matched the code will continue as known.
The function is no longer a static method for better access to the read patterns that are kept in a class variable.