🐛 Fix incidence calculation
There was some edge case handling which caused precision errors in the calculation.
Since this handling isn't needed anymore it gets removed.
It handled the case of gaps in the data, which actually don't occur since the introduction of requerster_v2.

Thanks to @thisryan for reviewing the code with me.

Co-authored-by: thisryan <oeyenadrian@gmail.com>