🆕 Add main function
The main in this file is useful for comparing the own calculated results (especially incidence) with those officially published.
It displays some arrays/ lists and diffs to compare data.

I also enabled the logging in one error handling.

Not a beauty but helpful :)

Thanks @thisryan for some minor bug fixes in the process.